v1.0.8.0
Change routing portion of the code so that the nodes will transmit two channel at once instead of one channel. This is due to discovery that on close proximity, power set above 50 will not work properly. 
Change 'B' to 'M' for the base in route table reporting.
Change the default configuration of freq (18, 28, 32, 36, 39, 43, 70, 80) and power of 350